coColorMap: fix for colormap tracking (maybe)
The transformMatrix have been computed by using cover->getViewerMat() * cover->getInvBaseMat(). For a viewer without headtracking it works fine, but with headtracking turned on the colormap will move to much around. Next try is to use the main camera view matrix instead to fix this issue. (Will be tested soon)